feat: product categories management

+++ product-service/src/main/java/dev/jakubdacewicz/product_service/category/dto/DetailedCategoryDto.java
@@ -12,7 +12,7 @@ public record DetailedCategoryDto(
         LocalDateTime createdAt,
         LocalDateTime updatedAt,
 
-        int productCount,
+        long productCount,
 
         boolean enabled) {
 }
